requirements.txt
Flask==2.3.3
Pillow==10.0.1
numpy==1.26.1
qrcode==7.4.2
qrcode[pil]
opencv-python==4.8.1.78